Add interface definition
