+++ src/Domain/Subscription/Betting/Persistence/BettingSubscriptionRepository.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace Statistico\Auth\Domain\Subscription\Betting\Persistence;
+
+use Ramsey\Uuid\UuidInterface;
+use Statistico\Auth\Domain\Subscription\BettingSubscription;
+use Statistico\Auth\Framework\Exception\NotFoundException;
+
+interface BettingSubscriptionRepository
+{
+    public function insert(BettingSubscription $subscription): void;
+
+    /**
+     * @param BettingSubscription $subscription
+     * @return void
+     * @throws NotFoundException
+     */
+    public function update(BettingSubscription $subscription): void;
+
+    /**
+     * @param UuidInterface $id
+     * @return BettingSubscription
+     * @throws NotFoundException
+     */
+    public function getByUserId(UuidInterface $id): BettingSubscription;
+
+    public function exists(UuidInterface $id): bool;
+}
